requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
nbconvert>=7.16.0
jupyter>=1.1.1
prophet>=1.1.5
Flask>=3.0.0
faiss-cpu>=1.8.0
pdfplumber>=0.11.0
requests>=2.32.0
beautifulsoup4>=4.12.0
